[refactor] adapter.in으로 파일 이동
